requirements.txt
# ETL
pandas==2.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0
faker==22.0.0

# OpenTelemetry
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-logging==0.43b0

# Datadog
ddtrace==2.6.0

# Metrics
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Logging
structlog==24.1.0
python-dotenv==1.0.0